Fixing test warning and move them to tests folder
